Structure is edited
