+++ backend/core/models.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+
+class BaseModel(models.Model):
+    id = models.AutoField(primary_key=True)
+
+    class Meta:
+        abstract = True
