+++ backend/npo_project/models/__init__.py
@@ -1,0 +1,2 @@
+from benefit import Benefit  # noqa
+from team_member import TeamMember  # noqa
