+++ backend/npo_project/models/benefit.py
@@ -1,31 +1,7 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-
-class BaseModel(models.Model):
-    id = models.AutoField(primary_key=True)
-
-    class Meta:
-        abstract = True
-
-
-class TeamMember(BaseModel):
-    name = models.CharField(
-        'Имя участника команды',
-        help_text='Введите имя, которое будет отображаться на сайте',
-        max_length=100,
-        unique=True
-    )
-    role = models.CharField('Роль в команде', max_length=100,)
-    image = models.ImageField('Фотография', upload_to='team_members/')
-
-    class Meta:
-        verbose_name = 'Участник команды'
-        verbose_name_plural = 'Участники команды'
-        ordering = ('name',)
-
-    def __str__(self):
-        return f'{self.name}, {self.role}'
+from core.models import BaseModel
 
 
 class Benefit(BaseModel):
