--- backend/npo_project/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render
-
-
-def index(request):
-    return render(request, 'index.html')
