+++ backend/setup.cfg
@@ -1,0 +1,37 @@
+[flake8]
+ignore =
+    #W503: Line break occurred before a binary operator
+    W503,
+    #I001: import in the wrong position
+    I001,
+    #I003: expected 1 blank line in imports, found 0
+    I003,
+    #I004: unexpected blank line in imports
+    I004,
+    #I005: unexpected missing import
+    I005
+exclude =
+    frontend/,
+    data/,
+    docs/,
+    */migrations/,
+    venv/,
+    env/,
+    infra/
+per-file-ignores =
+    */settings.py:E501
+max-complexity = 10
+
+[isort]
+skip=.git,.pytest-cache
+skip_glob=*/migrations/*,*/__pycache__/*,*/venv/*,
+use_parentheses = true
+balanced_wrapping = false
+default_section = THIRDPARTY
+force_single_line = false
+force_to_top = django
+include_trailing_comma = true
+known_third_party = django
+line_length = 79
+lines_between_types=1
+multi_line_output = 5
